+++ samples/download-svf.js
@@ -1,16 +1,17 @@
 /*
  * Example: downloading SVF assets for all viewables in a Model Derivative URN.
  * Usage:
- *     export FORGE_CLIENT_ID=<your client id>
- *     export FORGE_CLIENT_SECRET=<your client secret>
+ *     export APS_CLIENT_ID=<your client id>
+ *     export APS_CLIENT_SECRET=<your client secret>
  *     node download-svf.js <your model urn> <output folder>
  */
 
-const { SvfDownloader } = require('..');
-const { FORGE_CLIENT_ID, FORGE_CLIENT_SECRET } = process.env;
+const { SvfDownloader, TwoLeggedAuthenticationProvider } = require('..');
+const { APS_CLIENT_ID, APS_CLIENT_SECRET } = process.env;
 
 async function run(urn, outputDir = '.') {
-    const downloader = new SvfDownloader({ client_id: FORGE_CLIENT_ID, client_secret: FORGE_CLIENT_SECRET });
+    const authenticationProvider = new TwoLeggedAuthenticationProvider(APS_CLIENT_ID, APS_CLIENT_SECRET);
+    const downloader = new SvfDownloader(authenticationProvider);
     const download = downloader.download(urn, { outputDir, log: console.log });
     await download.ready;
 }