+++ samples/remote-svf-props.js
@@ -1,23 +1,21 @@
 /*
  * Example: parsing object properties for all viewables in a Model Derivative URN.
  * Usage:
- *     export FORGE_CLIENT_ID=<your client id>
- *     export FORGE_CLIENT_SECRET=<your client secret>
+ *     export APS_CLIENT_ID=<your client id>
+ *     export APS_CLIENT_SECRET=<your client secret>
  *     node remote-svf-props.js <your model urn>
  */
 
-const { ModelDerivativeClient, ManifestHelper } = require('forge-server-utils');
-const { SvfReader } = require('..');
+const { getSvfDerivatives } = require('./shared.js');
+const { SvfReader, TwoLeggedAuthenticationProvider } = require('..');
 
-const { FORGE_CLIENT_ID, FORGE_CLIENT_SECRET } = process.env;
+const { APS_CLIENT_ID, APS_CLIENT_SECRET } = process.env;
 
-async function run (urn) {
-    const auth = { client_id: FORGE_CLIENT_ID, client_secret: FORGE_CLIENT_SECRET };
-    const modelDerivativeClient = new ModelDerivativeClient(auth);
-    const helper = new ManifestHelper(await modelDerivativeClient.getManifest(urn));
-    const derivatives = helper.search({ type: 'resource', role: 'graphics' });
-    for (const derivative of derivatives.filter(d => d.mime === 'application/autodesk-svf')) {
-        const reader = await SvfReader.FromDerivativeService(urn, derivative.guid, auth);
+async function run(urn) {
+    const derivatives = await getSvfDerivatives(urn, APS_CLIENT_ID, APS_CLIENT_SECRET);
+    const authenticationProvider = new TwoLeggedAuthenticationProvider(APS_CLIENT_ID, APS_CLIENT_SECRET);
+    for (const derivative of derivatives) {
+        const reader = await SvfReader.FromDerivativeService(urn, derivative.guid, authenticationProvider);
         const propdb = await reader.getPropertyDb();
         const props = propdb.getProperties(1);
         for (const name of Object.keys(props)) {