+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [5.0.0] - 2024-04-09
+
+- Modified
+  - **[BREAKING CHANGE]** Library has been renamed from `forge-convert-utils` to `svf-utils`
+  - **[BREAKING CHANGE]** SVF readers and downloaders now expect an `IAuthenticationProvider` interface
+  for specifying how the requests to the Model Derivative service will be authenticated
+  - Changed branding from Forge to APS everywhere
+  - Migrated to the official APS SDKs
+
 ## [4.0.5] - 2023-09-29
 
 - Added